Added Module System and Autorization
I changed the server architecture to allow for the use of module base design.
The rest calls have been converted to a module.
A key manager has been added to start autorization. The keys will be loaded from a file so that we can remove all keys and tokens for 3rd party API's out of the repo.