Hooked up adding a new item to the rest server
I completed the rest request for adding a new item to the db. It needs work because there are no checks and all data is hard coded. An API for creaing items in both the angular front and node.js back end would make this much easier.

An example is in mappy-ctrl.js