Fixed server crash form bad endpoint data
I now check for a semi valid endpoint before trying to query the data.
I am now parsing the return data from the sql calls in a try catch block incase we have a bad response.
Restructured the data_manager.js to conform to the design paridigm.
Added a little icon for the tools program.